Removed old Header file including old Header.scss Updated new Header filepath
