--- printlib-frontend/src/components/layout/header.module.scss
@@ -1,84 +1,0 @@
-@use '../../styles/variables' as *;
-
-.header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem 2rem;
-    min-height: 10vh;
-    background-color: var(--color-primary-dark);
-}
-
-.logo {
-    height: 80px;
-    width: auto;
-}
-
-.nav {
-    display: flex;
-    gap: 1rem;
-}
-
-.nav a {
-    color: var(--color-border);
-    text-decoration: none;
-    font-weight: 500;
-    &:hover {
-        filter: brightness(1.8);
-    }
-}
-
-.nav a:hover {
-    text-decoration: underline;
-}
-
-.iconButton {
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding: 0.5rem;
-    font-size: 1.2rem;
-    color: none;
-}
-
-.menu {
-    position: relative;
-    display: inline-block;
-    height: 80px;
-    border-radius: 12px;
-    opacity: 50%;
-    &:hover {
-        opacity: 100%;
-        box-shadow: 0 4px 8px rgba(0, 0, 0, 0.5);
-    }
-    background: linear-gradient(to bottom, #ffffff 0%, #e0e0e0 100%);
-    border: 1px solid #aaa;
-}
-
-.middleGroup {
-    flex: .8;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 0 2rem;
-}
-
-.searchBar {
-    background-color: none;
-    color: var(--color-primary-dark);
-    padding: 0.5rem 1rem;
-    font-size: 1rem;
-    text-align: center;
-    border: none;
-    border-radius: $border-radius * 10;
-    width: 100%;
-    max-width: 1600px;
-    height: 40px;
-    opacity: 30%;
-    &:focus {
-        outline: none;
-        box-shadow: 0 4px 18px rgb(0, 0, 0);
-        border-color: var(--color-accent);
-        opacity: 50%;
-    }
-}
